Backend/requirements.txt
pandas
scikit-learn
joblib
fastapi
uvicorn
requests
python-dotenv
python-multipart